+++ cmd/fluxplane/main.go
@@ -1,0 +1,16 @@
+package main
+
+import (
+	"fmt"
+	"os"
+
+	fluxplaneapp "github.com/fluxplane/engine/apps/fluxplane"
+)
+
+func main() {
+	cmd := fluxplaneapp.NewCommand()
+	if err := cmd.Execute(); err != nil {
+		fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+}